CI: Test on Python 3.7 as well
According to the Travis-CI docs, this requires switching to a newer base
OS image. It might not require an explicit pin like this in near future
as the default is, apparently, being migrated at this very time, but I
do not feel like working with a random rolling-update process of a
third-party service, thank you.