docs: Add raw console capture of transmission-main-example.py
This is my source material of the animated picture which shows this tool
in action. Captured via [`asciinema`](https://asciinema.org/).

This includes changes from #261. I made a mistake by saying that the SNR
is in dBm instead of dB. Because this is a screencast in a plaintext
format, I was able to fix this without re-recording the demo, and that's
why I'm adding the source file now.