Correction: previus uploaded topo model was not the good one
upload the correct version of the file yangModels/gnpy-network-topology.yang
without the @ in the name. This one corresponds to the use in the TPCE
project https://git.opendaylight.org/gerrit/#/c/81785/
version of gnpy used for TPCE is here
https://github.com/Orange-OpenSource/gnpy/tree/api-rest-v0

the filename is the same but files are not the same.

Signed-off-by: EstherLerouzic <esther.lerouzic@orange.com>