Fix typos in Raman's frequency offset
This is a typo, the parameters were clearly supposed to be uniformly
spaced. It doesn't affect the results (much), because the remaining
values are "close enough" for a reasonable interpolation.

Using the existing Raman example, the difference in OSNR ASE at the
signal bandwidth is 0.01 dB (31.46 -> 31.47 dB).

Perhaps it would make sense to enforce that these offsets are a
monotonic sequence.