Unify transmission_main_example and the Raman wrapper
There were just these substantial differences:

- the Raman code showed a per-channel SNR summary, this is now
  controlled via the `--show-channels` option

- when a Raman fiber is used the `--sim` option for specifying input
  simulation parameters is now mandatory

I'm therefore merging these two files even though we've rpeviously
decided not to do this -- consult the review comment at
https://github.com/Telecominfraproject/oopt-gnpy/pull/263#discussion_r310506082
and the discussion during this Tuesday's coders call). If this turns out
to be a problem for autodesign, we can always revert this.

One possible catch is that the final "SNR total" shows NaN for the
default Raman example. That's just the way the simulation engine works
right now, I'm afraid. The `--show-channels` options helps a lot.