Update SNR before printing Transceiver output
This modifies the propagate functions in request.py so that the update_snr function of the last transceiver is called before printing the transceiver output (when show=True) in order to show the correct SNR
including tx_osnr and add_drop_osnr contributions. I could not come up with a better way than to split the last element out from the for loop and repeating some code.

Fixes #262