Enable ILA and FUSED type in routing constraints with xls input
here are the steps:
- builds correspondance dicts between input name from excel
  and names created with convert.py and autodesign in network.py
- correct the corresp_name dicts according to the effective
  network autodesign. This supports the case of fiber splitting
  and of fused elements
  (include the case of parrallel links with only one hop)
- interpret the node list constraint given by the user with the dict
- filter the constraints that are not applicable

This is needed because:

ILA and FUSED constraints must be filled with the next node
information in order to avoid confusion on the direction.
for example
     eg    a----b-----c
           |    |     |
           i    j     k
           |    |     |
           e----f-----g

a constraint 'j' given for service i to k leads to 2 possible direction:
i-a-b-j-f-g-k
i-e-f-j-b-c-k
the user must indicate the chosen direction. This ambiguity does not
exist with network input in json format (names are unique).

this commit is the first step:
it build a correspondance dictionary. For each city name (input in an
excel way) it gives the list of possible names in the topology.
for example user may enter foo name in node sheet, and the name in
the topology might be
- if foo is a roadm 'roadm foo', 'trx foo',
  + the preamp and booster for the roadm 'east edfa in foo to bar',
  'west edfa in foo to bar', 'Edfa0_fiber (foo → bar)-'
- if foo is an ILA 'east edfa in foo to bar', 'west edfa in foo to bar',
  'Edfa0_fiber (foo → bar)-',
- if foo is a fused 'west fused spans in foo'

When the user express foo as constraint, there my be ambiguity if node
has been defined as ila or fiber

The function corresp_names reflects what convert.py builds.

Signed-off-by: EstherLerouzic <esther.lerouzic@orange.com>