Define a function that correct constraint in case of excel service input
It corrects constraint list of nodes to be consistant with network nodes names
removes wrong names, find correct names for ila, roadm and fused if the entry was
xls.
if it was not xls, all names in list should be exact name in the network.
how it works:
-1- removes end transponders that do not need any correction
-2- check if the name is already in the list of correct nodes. if so, keep it as is
    else:
-3- checks if the name is part of correspondance keys. if so use the dict correspondance.
    the order is important first roadm, then fused and last ila
    at this point there is a list of node suggestions
-4- if there is more than one candidate, take the first one for which next node
    (using the function) is the next node in the constraint list
if no candidate satisfy the constraint and the consraint is strict, then stops the
simulation.
if constraint is loose, then ignore it, but warns the user that the constraint could
not be applied.

a series of test verify the consistancy of the constraint

Signed-off-by: EstherLerouzic <esther.lerouzic@orange.com>