correct test files because of new limitation
Here is the explanation for some of the corrections
Note1 without this work, the testTopology data file used for testing has a specified
gain (18.5) that leads to output power above max amplifier limit in ILA in St Brieuc
(ILA named east edfa in Stbrieuc to Rennes_STA). As a result the gain is cut to 13.18
instead of 18.5 during propagation. Since this does not affect the dp computation
during autodesign so there is still a 5 dB dp computed for the next edfa element which
explains the 6.5 dB gain in Edfa0_fiber (Stbrieuc → Rennes_STA)-F057
With this PR, the autodesign now sets 18.5 Gain (max 16 +2.5 exteded gain) to the
previous amplifier (east edfa in Lannion_CAS to Stbrieuc) (which is imposed to be
std_low_gain) instead of 20dB. This leads to 1.5 dB less in 'east edfa Stbrieuc to Rennes_STA'.
and the gain is thus cut to 13.18 + 1.5 = 14.68 dB. This 1.5 delta power is also reported
to the next elements so that the next edfa gain is set to 6.5 + 1.5 = 8 dB

TODO the gain cut that happen a posteriori should be included as a limitation during
the design.

Note 2 In the case of CORONET_Global_Topology data file changes are linked to
- python float numbers handling
- round2float function which limits the precision of float numbers to 1e-2

With this work, the computed gain is cut sharp to the max gain + extended gain so that
approx values are avoided.

Signed-off-by: EstherLerouzic <esther.lerouzic@orange.com>