Fix node_list and "STRICT" checking
The check on node_list constraint has to be be done on all elements except the last one,
and the last one is always a "STRICT" value. this means that if no constraints are given
this list is equals to ["STRICT"]. Previous code was using len("STRICT") ie 6 instead of
len(["STRICT"]) ie 1.

Signed-off-by: EstherLerouzic <esther.lerouzic@orange.com>