Zuul: test with Python 3.6 as well
We're saying that Python 3.6 is the minimal version that we support, so
let's make sure there's CI coverage for that.

I also tried enabling Python 3.8, but somehow the build of Pandas
failed. I don't feel like debugging a Pandas build failure today, so
let's postpone this thing until later. Just having 3.6 is a net
improvement, and we can play with even newer Python later -- and perhaps
on a newer distro, anyway.

Change-Id: I28a8c282225b7070ed3dddba56cccc8def313a77