CI: build documentation, too
I've bumped Sphinx because this might be needed for proper support of
pbr-provided version numbers. In the end I also had to re-run `pip
install -e .` locally, quite likely due to a different format of version
numbers now that pbr is being used, but let's make sure that we're using
a version of Sphinx that is now known to work.

Change-Id: I451b4b17bb8097eb7f2f0f0956cc1c956a531828
Depends-On: https://review.opendev.org/715028