Update patch set 1
Patch Set 1: Code-Review-1

I have no clue whether this is correct or not, and a quick Google search did not lead to a definition of fiber gamma. Please either self-approve this, or let someone from Polito do that, or (preferably in a commit message) point me to a paper which explains this to me.

Patch-set: 1
Reviewer: Gerrit User 1009907 <1009907@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1