Update patch set 1
Patch Set 1:

Thanks for the review. I understand your concern, but I do not agree with you. Being able to support ILA in constraints is mandatory in order to support parallel links, and our user mainly use excel. So working on this excel input is important and should not add complexity to the user. Let's discuss this during our next call.

Patch-set: 1